requirements.txt
requests
beautifulsoup4
pandas
matplotlib
seaborn
plotly
dash
flask
apscheduler
python-dotenv
email-validator
